requirements.txt
google-cloud-storage
google-cloud-datastore  
google-cloud-secret-manager
psycopg2-binary
numpy>=1.19.0
pandas
xlrd
pyarrow
jsonschema
altair
altair_saver
pyyaml
flask
gunicorn
streamlit
mysqlclient
more_itertools
sqlalchemy
Deprecated